phase1/requirements.txt
# Core data processing
pandas>=2.0.0
numpy>=1.24.0

# Visualisation
matplotlib>=3.7.0

# Google Cloud – Storage (read NGSIM CSV from GCS)
google-cloud-storage>=2.10.0

# Google Cloud – BigQuery (write scenario samples)
google-cloud-bigquery>=3.11.0

# Required by BigQuery load_table_from_dataframe
pyarrow>=12.0.0
